Ship tested GameCube iOS performance build
Add modern iPhone and iPad import, landscape, JIT, Metal and performance fixes together with reproducible simulator, packaging, public-audit, and agent bootstrap workflows.

+++ Source/Core/InputCommon/ControlReference/ControlReference.h
@@ -72,7 +72,7 @@ inline bool ControlStateCast<bool>(ControlState value)
 template <>
 inline int ControlStateCast<int>(ControlState value)
 {
-  return std::lround(value);
+  return static_cast<int>(std::lround(value));
 }
 
 template <>
